pyproject.toml
[tool.poetry]
name = "nonebot_mockingbird_plugin"
version = "0.0.2"
description = "A Nonebot Use Version for MockingBird"
authors = ["Alex Newton <sharenfan222@gmail.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/babysor/MockingBird"
repository = "https://github.com/Diaosi1111/nonebot_mockingbird_plugin"
keywords = ["MockingBird", "python"]

[tool.poetry.dependencies]
python = "^3.9"
MockingBirdOnlyForUse = "*"
nonebot2="^2.0.0b1"

[tool.poetry.dev-dependencies]

[tool.poetry.extras]

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
src_paths = ["nonebot", "tests"]
extra_standard_library = ["typing_extensions"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"